4.21*: bloglist expansion, step9

+++ part_4/controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require("../models/blog")
 const User = require("../models/user")
 const jwt = require("jsonwebtoken")
 
-// const getTokenFrom = request => {
-//   const authorization = request.get("authorization")
-//   if (authorization && authorization.startsWith("Bearer ")) {
-//     return authorization.replace("Bearer ", "")
-//   }
-//   return null
-// }
-
 blogsRouter.get("/", async(request, response) => {
   const blogs = await Blog
     .find({}).populate("user", { username: 1, name: 1 })
@@ -18,10 +10,17 @@ blogsRouter.get("/", async(request, response) => {
   response.json(blogs)
 })
 
+blogsRouter.get("/:id", async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog)
+  } else {
+    response.status(404).end()
+  }
+})
+
 blogsRouter.post("/", async (request, response) => {
-  console.log("request in blogsRouter post is", request)
   const body = request.body
-  console.log("request.token is", request.token)
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token invalid" })
@@ -49,8 +48,14 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(400).json({ error: "wrong user" })
+  }
 })
 
 blogsRouter.put("/:id", async (request, response) => {
